pyproject.toml
[project]
name = "bimsa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "e2cnn>=0.2.3",
    "einops>=0.8.1",
    "flashtorch>=0.1.3",
    "ipykernel>=6.30.1",
    "jax>=0.5.3",
    "jaxtyping>=0.3.2",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.3",
    "nose>=1.3.7",
    "numba>=0.61.2",
    "numpy==1.26",
    "pandas>=2.3.2",
    "pyinstrument>=5.1.2",
    "pyseagull>=1.0.0b4",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "setuptools>=82.0.1",
    "sympy>=1.14.0",
    "thop>=0.1.1.post2209072238",
    "toml>=0.10.2",
    "torch>=2.10.0",
    "torch-geometric>=2.7.0",
    "torchinfo>=1.8.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.1",
    "wandb>=0.21.0",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[[tool.uv.index]]
name = "pyg-cu130"
url = "https://data.pyg.org/whl/torch-2.10.0+cu130.html"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu130", marker = "sys_platform == 'win32'" },
]
pyg-lib = [
  { index = "pyg-cu130", marker = "sys_platform == 'win32'" },
]
